# Documents/openapi.yaml
---
openapi: 3.0.0
info:
  title: URL shortener API
  description: URL shortener
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0-oas3
servers:
- url: https://virtserver.swaggerhub.com/jorgeRambla/urlShortener/1.0.0
- url: http://virtserver.swaggerhub.com/jorgeRambla/urlShortener/1.0.0
tags:
- name: main
  description: main operations
- name: redirect
  description: redirect operations
- name: stats
  description: stats operations
paths:
  /short:
    post:
      tags:
      - main
      summary: Create new shortend URL
      operationId: postCreateShortenedURL
      parameters:
      - name: headURL
        in: query
        description: URL to be shortened
        required: true
        style: FORM
        explode: true
        schema:
          type: string
      - name: intersticialURL
        in: query
        description: Intersticial URL
        required: false
        style: FORM
        explode: true
        schema:
          type: string
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/URL'
        400:
          description: bad input parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /redirect/{sequence}:
    get:
      tags:
      - redirect
      summary: Get shortened URL, updating stats
      operationId: getShortenedURL
      parameters:
      - name: sequence
        in: path
        description: Shortened URL sequence code
        required: true
        style: SIMPLE
        explode: false
        schema:
          type: string
      - name: userAgent
        in: query
        description: UserAgent, stats to update
        required: false
        style: FORM
        explode: true
        schema:
          type: string
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedirectURL'
        400:
          description: bad input parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /qr/{sequence}:
    get:
      tags:
      - main
      summary: Get the QR code of the shortened URL
      operationId: getQR
      parameters:
      - name: sequence
        in: path
        description: Shortened URL sequence code
        required: true
        style: SIMPLE
        explode: false
        schema:
          type: string
      responses:
        201:
          description: Success
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/svg+xml:
              schema:
                type: string
        400:
          description: bad input parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /stats/daily/{sequence}:
    get:
      tags:
      - stats
      summary: Get the stats of the shortened URL
      operationId: getStats
      parameters:
      - name: sequence
        in: path
        description: Shortened URL sequence code
        required: true
        style: SIMPLE
        explode: false
        schema:
          type: string
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stats'
        400:
          description: bad input parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /stats/global/{sequence}:
    get:
      tags:
      - stats
      summary: Get the stats of the shortened URL
      operationId: getGlobalStats
      parameters:
      - name: sequence
        in: path
        description: Shortened URL sequence code
        required: true
        style: SIMPLE
        explode: false
        schema:
          type: string
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClickStat'
        400:
          description: bad input parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /browsers:
    get:
      tags:
      - main
      summary: Get supperted browsers
      operationId: getSupportedBrowsers
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Browser'
components:
  schemas:
    URL:
      required:
      - URL
      type: object
      properties:
        URL:
          type: string
          example: https://www.google.es
    ClickStat:
      required:
      - browser
      - clicks
      type: object
      properties:
        browser:
          type: string
          example: Google Chrome
        clicks:
          type: string
          format: int32
          example: "154"
    Stats:
      required:
      - clickStat
      - day
      type: object
      properties:
        day:
          type: string
          format: date
        clickStat:
          type: array
          items:
            $ref: '#/components/schemas/ClickStat'
    RedirectURL:
      required:
      - headURL
      - interstitialURL
      type: object
      properties:
        headURL:
          type: string
          example: https://www.google.es
        interstitialURL:
          type: string
          example: https://www.unizar.es
    Error:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: string
          format: int32
          example: "404"
        message:
          type: string
          example: Shortened URL not found
    Browser:
      required:
      - browser
      type: object
      properties:
        browser:
          type: string
          example: Opera
  extensions: {}