# spring-server-generated/src/main/java/io/swagger/api/QrApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.2).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.Error;
import java.io.File;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2018-11-12T11:58:42.466Z[GMT]")

@Api(value = "qr", description = "the qr API")
public interface QrApi {

    @ApiOperation(value = "Get the QR code of the shortened URL", nickname = "getQR", notes = "", response = File.class, tags={ "main", })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "Success", response = File.class),
        @ApiResponse(code = 400, message = "bad input parameter", response = Error.class),
        @ApiResponse(code = 404, message = "not found", response = Error.class) })
    @RequestMapping(value = "/qr/{sequence}",
        produces = { "image/png", "image/svg+xml", "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<File> getQR(@ApiParam(value = "Shortened URL sequence code",required=true) @PathVariable("sequence") String sequence);

}