# spring-server-generated/src/main/java/io/swagger/api/RedirectApiController.java
package io.swagger.api;

import io.swagger.model.Error;
import io.swagger.model.RedirectURL;
import com.fasterxml.jackson.databind.ObjectMapper;
import io.swagger.annotations.*;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.constraints.*;
import javax.validation.Valid;
import javax.servlet.http.HttpServletRequest;
import java.io.IOException;
import java.util.List;
import java.util.Map;
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2018-11-12T11:58:42.466Z[GMT]")

@Controller
public class RedirectApiController implements RedirectApi {

    private static final Logger log = LoggerFactory.getLogger(RedirectApiController.class);

    private final ObjectMapper objectMapper;

    private final HttpServletRequest request;

    @org.springframework.beans.factory.annotation.Autowired
    public RedirectApiController(ObjectMapper objectMapper, HttpServletRequest request) {
        this.objectMapper = objectMapper;
        this.request = request;
    }

    public ResponseEntity<RedirectURL> getShortenedURL(@ApiParam(value = "Shortened URL sequence code",required=true) @PathVariable("sequence") String sequence,@ApiParam(value = "UserAgent, stats to update") @Valid @RequestParam(value = "userAgent", required = false) String userAgent) {
        String accept = request.getHeader("Accept");
        return new ResponseEntity<RedirectURL>(HttpStatus.NOT_IMPLEMENTED);
    }

}